fix: token retrieval and UTC datetime cookie error
The token dependency injection returned a decoded cookie, which worked
fine for the SignUp endpoint as it can retrieve the email without an
extra step. However, to create a session cookie, a raw token is required,
so passing the decoded token from the dependency led to an invalid token
error.

Now, the token dependency returns a raw token. So the login endpoint now
turns it into a decoded token as an extra step. The signup endpoint passes
the token to cookie function as is.

Also, the expiration datetime for cookies must be in UTC format so that
was also fixed.