feat: admin router and delete post by id
Created a router for admin endpoints.

To check a if a user is an admin, a new dependency was created that uses
the get_user_by_cookie function to first get the current user (if
they're signed in and a session cookie exists). Then, the user's is_admin
status is checked. If they are an admin, we return the current_user. If
not, an HTTP exception is raised.

Every endpoint in admin will use this dependency so that it checks if a
user is signed in AND if the user has admin privileges.

Finally, connected the routers in main.py