fix: use unique vehicles dataset on start up
If we used the original dataset and eliminated rows in real time, this
would take about a minute if it ran locally. However, when deployed on
Render, it takes too long and the vehicle import gets cancelled
prematurely because the backend spins down after about 15 min.

This change now uses a vehicle dataset that has already had duplicate
vehicles removed to improve import speed.